+++ src/config/userRules.js
@@ -70,6 +70,7 @@ export const rulesForUserRoles = {
       rulesList.dealer_credit_view,
       rulesList.dealership_credit_view,
       rulesList.credit_view,
+      rulesList.pdr_view,
       rulesList.financial_view,
       rulesList.upload_statement,
     ]
@@ -101,6 +102,7 @@ export const rulesForUserRoles = {
       rulesList.dealer_credit_edit,
       rulesList.dealership_credit_edit,
       rulesList.credit_view,
+      rulesList.pdr_view,
       rulesList.financial_view,
       rulesList.upload_statement,
     ]
@@ -135,6 +137,7 @@ export const rulesForUserRoles = {
       rulesList.dealer_credit_edit,
       rulesList.dealership_credit_edit,
       rulesList.credit_view,
+      rulesList.pdr_view,
       rulesList.financial_view,
       rulesList.upload_statement,
     ]