+++ src/pages/reports/CreditNewRequestTable.js
@@ -12,18 +12,13 @@ import {
   getTypeOfAccount,
 } from '../../services/users.service';
 
-
 const CreditNewRequestTable = ({ data, currentUser, view }) => {
-  const [tableData, setTableData] = useState([]);
-  const [processedData, setProcessedData] = useState([]);
   const [accountType, setAccountType] = useState();
   const [rowData, setRowData] = useState();
   const [loading, setLoading] = useState(false);
   const [openModal, setOpenModal] = useState(false);
   const [statusModal, setStatusModal] = useState(false);
-  const [selectedTab, setSelectedTab] = useState('processed');
-  const [dealershipData, setDealershipData] = useState();
-
+  usePageTitle('Credit Reload');
 
   useMount(() => {
     getTypeOfAccount()
@@ -41,7 +36,7 @@ const CreditNewRequestTable = ({ data, currentUser, view }) => {
         console.log(e);
       })
   });
-  usePageTitle('Credit Report');
+
   const columns = useMemo(() => {
     return [
       {
@@ -191,7 +186,6 @@ const CreditNewRequestTable = ({ data, currentUser, view }) => {
           <CreditReloadForm
             callback={() => setOpenModal(false)}
             data={accountType}
-            dealershipData={dealershipData}
             currentUser={currentUser}
             view={view}
           />